usm: Unify eBPFProgram and USM Monitor
Over the time, there was no clear separation between USM monitor and the eBPF program struct. We ended up with having a mix, and complex code
that access the internal fields of the eBPFProgram struct.
