--- pkg/network/usm/dump.go
@@ -1,85 +1,0 @@
-// Unless explicitly stated otherwise all files in this repository are licensed
-// under the Apache License Version 2.0.
-// This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2016-present Datadog, Inc.
-
-//go:build linux_bpf
-
-package usm
-
-import (
-	"strings"
-	"unsafe"
-
-	manager "github.com/DataDog/ebpf-manager"
-	"github.com/cilium/ebpf"
-	"github.com/davecgh/go-spew/spew"
-
-	"github.com/DataDog/datadog-agent/pkg/network/protocols/http"
-)
-
-func (e *ebpfProgram) dumpMapsHandler(manager *manager.Manager, mapName string, currentMap *ebpf.Map) string {
-	var output strings.Builder
-
-	switch mapName {
-	case sslSockByCtxMap: // maps/ssl_sock_by_ctx (BPF_MAP_TYPE_HASH), key uintptr // C.void *, value C.ssl_sock_t
-		output.WriteString("Map: '" + mapName + "', key: 'uintptr // C.void *', value: 'C.ssl_sock_t'\n")
-		iter := currentMap.Iterate()
-		var key uintptr // C.void *
-		var value http.SslSock
-		for iter.Next(unsafe.Pointer(&key), unsafe.Pointer(&value)) {
-			output.WriteString(spew.Sdump(key, value))
-		}
-
-	case "ssl_read_args": // maps/ssl_read_args (BPF_MAP_TYPE_HASH), key C.__u64, value C.ssl_read_args_t
-		output.WriteString("Map: '" + mapName + "', key: 'C.__u64', value: 'C.ssl_read_args_t'\n")
-		iter := currentMap.Iterate()
-		var key uint64
-		var value http.SslReadArgs
-		for iter.Next(unsafe.Pointer(&key), unsafe.Pointer(&value)) {
-			output.WriteString(spew.Sdump(key, value))
-		}
-
-	case "bio_new_socket_args": // maps/bio_new_socket_args (BPF_MAP_TYPE_HASH), key C.__u64, value C.__u32
-		output.WriteString("Map: '" + mapName + "', key: 'C.__u64', value: 'C.__u32'\n")
-		iter := currentMap.Iterate()
-		var key uint64
-		var value uint32
-		for iter.Next(unsafe.Pointer(&key), unsafe.Pointer(&value)) {
-			output.WriteString(spew.Sdump(key, value))
-		}
-
-	case "fd_by_ssl_bio": // maps/fd_by_ssl_bio (BPF_MAP_TYPE_HASH), key C.__u32, value uintptr // C.void *
-		output.WriteString("Map: '" + mapName + "', key: 'C.__u32', value: 'uintptr // C.void *'\n")
-		iter := currentMap.Iterate()
-		var key uint32
-		var value uintptr // C.void *
-		for iter.Next(unsafe.Pointer(&key), unsafe.Pointer(&value)) {
-			output.WriteString(spew.Sdump(key, value))
-		}
-
-	case "ssl_ctx_by_pid_tgid": // maps/ssl_ctx_by_pid_tgid (BPF_MAP_TYPE_HASH), key C.__u64, value uintptr // C.void *
-		output.WriteString("Map: '" + mapName + "', key: 'C.__u64', value: 'uintptr // C.void *'\n")
-		iter := currentMap.Iterate()
-		var key uint64
-		var value uintptr // C.void *
-		for iter.Next(unsafe.Pointer(&key), unsafe.Pointer(&value)) {
-			output.WriteString(spew.Sdump(key, value))
-		}
-
-	case connectionStatesMap: // maps/connection_states (BPF_MAP_TYPE_HASH), key C.conn_tuple_t, value C.__u32
-		output.WriteString("Map: '" + mapName + "', key: 'C.conn_tuple_t', value: 'C.__u32'\n")
-		iter := currentMap.Iterate()
-		var key http.ConnTuple
-		var value uint32
-		for iter.Next(unsafe.Pointer(&key), unsafe.Pointer(&value)) {
-			output.WriteString(spew.Sdump(key, value))
-		}
-
-	default: // Go through enabled protocols in case one of them now how to handle the current map
-		for _, p := range e.enabledProtocols {
-			p.DumpMaps(&output, mapName, currentMap)
-		}
-	}
-	return output.String()
-}
